refactor(sso): share access token payload type across clients
